Update s3 factory class to have a default env of development.
