+++ src/lib/S3-factory.ts
@@ -1,8 +1,8 @@
 import { S3 } from 'aws-sdk';
 
 export default class s3Factory {
-  public static getS3Instance(environment: string = 'test', email: string) {
-    if (environment.toLowerCase() == 'test') {
+  public static getS3Instance(config = {env: 'development'}) {
+    if (config.env.toLowerCase() == 'test') {
       let s3Object = new S3();
       let sinon = require('sinon');
       sinon.stub(s3Object, 'putObject');
