--- test/support/AWS-wrapper.ts
@@ -1,17 +1,0 @@
-import {S3} from 'aws-sdk';
-import s3Factory  from './S3-factory'; 
-
-export class FileStorage {
-  _S3: S3;
-  email: string;
-  constructor(email: string) {
-    this._S3 = s3Factory.getS3Instance('development', email);
-    this.email = email;
-  }
-
-  get S3() { return this._S3; }
-
-  upload() {
-    this.S3.putObject()
-  }
-}
